requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.3.0
distlib==0.3.6
filelock==3.12.2
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
pipenv==2023.7.4
platformdirs==3.8.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.23.1
virtualenv-clone==0.5.7
